fixed later fuel to float

+++ database/migrations/2017_02_12_074143_create_fuels_table.php
@@ -23,7 +23,7 @@ class CreateFuelsTable extends Migration
             $table->string('gas_type');
             $table->integer('last_number_car');
             $table->integer('present_number_car');
-            $table->integer('liter');
+            $table->float('liter', 10, 2);
             $table->integer('total_distance');
             $table->float('gas_per_distance', 10, 2);
             $table->mediumText('note')->nullable();
